stripe wants ints

+++ payment/views.py
@@ -34,7 +34,7 @@ def payment_process(request):
         for item in order.items.all():
             session_data['line_items'].append({
                 'price_data': {
-                    'unit_amount': item.price * Decimal('100'),
+                    'unit_amount': int(item.price * Decimal('100')),
                     'currency': 'brl',
                     'product_data': {
                         'name': item.product.name,
